Move 'onClickSave' processing logic from EmailSender to ManageSender
